requirements.txt
django-jsonfield==0.9.13

Pillow==2.7.0

redis==2.10.3

pycurl==7.19.5

requests==2.5.1

django-jsonview==0.4.3

alligator==0.8.0

Django==1.7.2

# TarGZ
# sha256: 8uJz7TSsu1YJYtXPEpF5NtjfAil98JvTCJuFRtRYQTg
# Wheel
# sha256: ygF2j97N4TQwHzFwdDIm9g7f9cOTXxJDc3jr2RFQY1M
dj-database-url==0.3.0

# sha256: JLo_Moq6CNjVBbwqj2ipTr4kqF6Eg2QLZnIhvA79Ox4
psycopg2==2.5.4

# sha256: ufv3R_qOcR0zDy9Da5suz17tnrZ6Y3yoGrzRHCq-3sg
python-decouple==2.3

# sha256: LiSsXQBNtXFJdqBKwOgMbfbkfpjDVMssDYL4h51Pj9s
Jinja2==2.7.3

# sha256: o71QUd0X6dUOip8tNiUI6C1wYwEP3AbV4HzjYI54D-Y
jingo==0.7

# sha256: pOwa_1m5WhS0XrLiN2GgF56YMZ2lp-t2tW6ozce4ccM
MarkupSafe==0.23

# TarGZ
# sha256: gnFVEftiRvrUumbYEuuTQWroNxtGT6iL84Z8nBd9qhQ
# Wheel
# sha256: _7QRdX3jdNBcdZySmQjSOXxm47z2813uQabvnrhRSXs
gunicorn==19.1.1

# TarGZ
# sha256: pNxeBMs75dG-9lGXFXl4Md77FtMB5193l85wxnyeNbo
# Wheel
# sha256: AQYO48CTYIdztcrXXioWrKd5GJDyQQgFOsNNB7Sd4Uk
whitenoise==1.0.4

# sha256: 0yV2uq13_dnmhZvpfCtqPBem9IQ2Q4nbcx8Gl7ehlmE
django-csp==2.0.3

# sha256: xby9Ec-YRwlq4etOg93nXRCsYu_m5zxGAPP5gJaM29I
# sha256: Qxx4vPgnySoZ14Kc2a55AtUtbfyyPZiQT8gH3eoewHY
# sha256: S6CrdPp_NPd_8jbuIVdFTRwqrqqiWTI2O_ob5suC29Y
pytz==2014.9

# sha256: drxjpOLV5aDfd8p9GPD1bixGz7YrcRA7qSqSx5-rHgM
nose==1.3.4

# sha256: mq4WtWKGak3apeiXhymrrbvtVEco2I4LnJr3sx3Qcu0
django-nose==1.2

# sha256: l87FJeejaziLKl9qV6MwWLIdrHTqpEtWpYLPIv2FuRI
newrelic==2.38.0.31

# sha256: KF6L1zDAtv374jwy0pNr_7pAHyPKsTLocixovoDW8YI
flake8==2.2.5

# Wheel Python 2.7
# sha256: q409VanEtW-Xy2D21rk92UlUwWEU8K6HLCR6zoI_Fds
# Wheel Python 3
# sha256: jy0vup7228b_a2mm76_Yg1voKHm4rnmkM2HH1k42Pew
# TarGZ
# sha256: PphGaqL-VFQLy6mqbgGjn0ARDWdmjClzQMS5UUt8xJw
# ZIP
# sha256: MmP8nqIHC-HP-ERXhtbzGLdnaMAGg4n7OFOw0us_THI
ipython==2.3.1

# sha256: DUoz845megyA7BpaDkaP5Y5c3GHLj0B3sQJLhf57cRc
ipdb==0.8

# Wheel Python 2
# sha256: 0UXkOs2Dxbos77J2EH6Wb-YzXbAmWnoLSxCirbhM0m0
# Wheel Python 3
# sha256: uwkN7xH7pRFnpYjYyklp9Jo8Q3rYrPlg39OcSwzxYAk
# TarGZ
# sha256: lJM7ZOL-CAfaBhLFdKAhwNrCjHvTxKI3I65aOeqPPQQ
Sphinx==1.2.3

# sha256: PYBK7mdyHzjoQO-jo65SiYMo5kcJJZNB8qiQHVvwHNQ
# sha256: Dyn1RPbQN5ifoMdymp6rfk2OpQ1vDvNzY_RydWwe3KY
sphinx-rtd-theme==0.1.6

# sha256: XinOsvZKY-HOMBr0x0vrJ7WmYeWGPQBFT1mUC6JNc50
# sha256: OQDKtkuxi2YYkSgARVAePxoFoeyZp1rrLtAmkdB4JcI
sphinx-autobuild==0.3.0

# sha256: x9txeBCraWX2bIzwOYqYydjfmC2jm0zX8WKRHriVlvo
docutils==0.12

# sha256: XgOeHUDSMpge1YkUttGsLkU6foPd6iLvnz7q3QHeRcs
# sha256: ZoyOZluTX2I4WcEyE-G8mzP9Gf7sTbfC-OhVwcjFWdc
# sha256: fydQ9H4cA_ffSGbTdhm4B0_8HTIYI7taNRMC_k3SC2c
Pygments==2.0.1
